Load and store RCSB metadata

+++ src/file_io/map.rs
@@ -18,6 +18,7 @@ use crate::reflection::ElectronDensity;
 const HEADER_SIZE: u64 = 1_024;
 
 /// Minimal subset of the 1024-byte CCP4/MRC header
+#[allow(unused)]
 #[derive(Clone, Debug)]
 pub struct MapHeader {
     pub nx: i32,      // grid points along X
@@ -338,7 +339,7 @@ pub fn read_map_data(path: &Path) -> io::Result<(MapHeader, Vec<ElectronDensity>
 
                 densities.push(ElectronDensity {
                     coords: coord,
-                    density: raw as f64,
+                    density: raw,
                 });
             }
         }
